requirements.txt
numpy
pandas
scikit-learn
jupyter
matplotlib
seaborn
mlflow
locust
kafka-python
setuptools
confluent-kafka
river
gevent
